+++ ZoomSDK/ZoomSDK.framework/Versions/A/Headers/ZoomSDKBuildType.h
@@ -1,5 +1,0 @@
-#pragma once 
-
-#define BUILD_FOR_ELECTRON 
-#define BUILD_FOR_SDK 
-